+++ src/event/win/Window.cpp
@@ -1,7 +1,8 @@
 #include "event/Window.hpp"
+#include "event/Input.hpp"
 #include <windows.h>
 
-int32_t OsWindowProc(void* window, uint32_t message, uint32_t wparam, int32_t lparam) {
+int32_t OsWindowProc(void* window, uint32_t message, uintptr_t wparam, intptr_t lparam) {
     // TODO
     return 0;
 }
